requirements.txt
PyQt5
SQLAlchemy
pyodbc
python-dotenv